+++ src/build-info.ts
@@ -1,2 +1,2 @@
-export const BUILD_LABEL = "ThemeDeck Build 2026-06-30 11:24:38 EDT";
-export const BUILD_ID = "2026-06-30-112438-EDT";
+export const BUILD_LABEL = "ThemeDeck Build 2026-09-06 11:44:23 EDT";
+export const BUILD_ID = "2026-09-06-114423-EDT";